requirements.txt
streamlit==1.39.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.28
langchain-core==0.3.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
typing-extensions==4.12.2
openai>=1.52.0
httpx>=0.27.0,<0.29.0
